refactoring log
