+++ src/pages/admin/techSinglePage/TechSinglePage.tsx
@@ -2,20 +2,7 @@ import { FC, useState } from 'react';
 import EditIcon from '@mui/icons-material/Edit';
 import { Box, Button, Container, Divider, Grid, TextField, Typography } from '@mui/material';
 
-const boxLogStyle = {
-    boxSizing: 'border-box',
-    display: 'flex',
-    flexDirection: 'column',
-    width: '100%',
-    bgcolor: 'background.paper',
-    position: 'relative',
-    overflow: 'auto',
-    maxHeight: '95%',
-    borderRadius: '5px',
-    padding: '15px 15px',
-    border: '1px solid',
-    borderColor: 'rgba(0, 0, 0, 0.23)',
-};
+import LogList from '../components/logList/LogList';
 
 const TechSinglePage: FC = () => {
     const [textAboutReadOnly, setTextAboutReadOnly] = useState<boolean>(true);
@@ -23,7 +10,6 @@ const TechSinglePage: FC = () => {
     return (
         <Container maxWidth="xl">
             <Box sx={{ display: 'flex' }}>
-                <input></input>
                 <Typography variant="h5" sx={{ textAlign: 'left', margin: '15px 0 15px 40px' }}>
                     Название технологии
                 </Typography>
@@ -103,20 +89,7 @@ const TechSinglePage: FC = () => {
                         maxHeight: '80vh',
                     }}
                 >
-                    <Typography variant="h5" sx={{ textAlign: 'center', height: '36px' }}>
-                        Лог изменений
-                    </Typography>
-                    <Box sx={boxLogStyle}>
-                        {[1, 1, 1, 1, 1, 1, 1, 1, 1, 1].map((_, index) => (
-                            <Button key={index} sx={{ borderBottom: '1px solid', marginBottom: '10px' }}>
-                                <Typography>
-                                    <strong>Дата: 11/04/2023 Время: 20:22 Радар: Frontend</strong>
-                                    <br />
-                                    Добавил преимущество, изменил ссылку на доку
-                                </Typography>
-                            </Button>
-                        ))}
-                    </Box>
+                    <LogList boxWidth="100%" />
                 </Grid>
             </Grid>
         </Container>
